Move all dataflow dependencies to accessors and remove dataflow utils
